Added - Portuguese Translation to logistics_uavbattery
The correct translation of "UAV" to Portuguese is indeed "VANT", but this term is not used as "UAV" is. In reality, most Portuguese speaking people knows what an "UAV" is but if you use the Portuguese translation they have no idea of what you are talking about.

Another alternative is to use the word "Drone". Everyone knows what a drone its, however, I find that the "UAV" is more in terms with the authenticity aspect of ACE3, since it the military uses the word "UAV" instead of "Drone"


Also, it seems that somebody wrote  "UAV's" instead of "UAVs". If you intentd to use the plural, you should not use the Apostrophe. I am not correcting this since I am not sure if indeed this was intended for plural or not.