French translation pass
some corrections and the medical stringtable completed:.

Be advised that i'm no medic and that some medical words "may" be wrong
in the context they are used, especially for velocity wounds. for the
rest it should be fine.

other changes:
claymore name changed, was too long
Sanitaire changed to Médical for trucks.

Some translation on missile guidance.

typo

typo

typo

space fixed for the claymore name

typos fixed

everything pointed out in the pull request