Create missing documentation placeholders
Creates placeholder documentation for modules which did not previously
have any. If a module had the same name and apparent purpose as a module
from AGM, AGM's documentation was imported into the placeholder.

These documentation are useful to users for explaining what ACE3
changes. They are also useful to community admins who want to create a
custom distribution of ACE3 and need to know which modules to disable.

The documentation needs to be significantly fleshed out before it's
ready for public consumption, but this should be a helpful start.