separating isCarrying/isDragging from startCarrying/startDragging isCarrying was used even when the start wasn't finished yet. While carried object is needed to abort anim, I made carrying/dragging more equal.
Thisfor, instead of is* for the beginning, start* will be used.
is* will be true once the caller is effectively carrying/dragging the target.

(I needed to know when the unit finished with the startCarry for AI to know when to go on.
isBeingDragged worked already correct, while isBeingCarried was incorrect).

Notice: is the Timer for Weapon for abortion really needed?
Because if you start carrying a unit, but abort it by switching to pistol, and then quickly drag the unit,
you will notice that after the timeOut you will lose the dragging Task.