Reset captive status on respawn for reliability
Previously we were resetting on death, but this caused problems when the respawnDelay was 0 as the dead unit wasn't actually the player.

To account for "group" and "side" respawn types where you change unit on respawn and could potentially become a captive unit, I am explicity checking the mission respawn type before resetting variables.