Improved efficiency of wound counting.
Using a switch/case conditional since it offers early termination if a
matching condition is found before all conditions are checked. It also
unnecessary to use a conditional to confirm whether a limb is wounded or
not. Number of Wounds multiplied by Blood loss will return zero for
a fully bandaged body part, not incrementing the wound counter; or it will
return some other number which will increment the wound counter.