Add initial explination of project goal system
As was discussed in the development meeting of 12/12/2015, we're implementing a new system of setting a specific goal to work towards between versions (dubbed 'project goals'). This is not a commitment of any kind, but rather a means of providing direction for collaborative development to focus in.

Initial explanation aims to make the intentions of the system clear and eliminate any potential for false expectations.