Remove headless client component from ACE.
The component does not fit properly within the scope of the ACE project. In order to avoid difficulties at a later point in time, project management has decided to pull the HC component from the ACE 3.5.0 release. We are going to look into alternative options for the HC component, such as a separate repository maintained by the ACE team.