Cleanup zeus display modules
Replaces display size and position defines with macros for readability and compactness.

Changes naming convention of functions and displays to distinguish categories and make more readable.

Converts the global set skill module to use display events rather than waiting in a loop for the module to be set and confirmed.