Add ammo cookoff (#4376)
* Add Ammo cookoff

* Remove tabs

* Add initial ammo box cook-off

Does not include a fire effect, mostly just a proof of concept.

Should probably also add further potential cook-off conditons (if hit by tracer for example).

* Add burning effects to ammo box cook off

- Add burning effect while ammo box is cooking off
- Add setting to enable/disable ammo boxes cooking off
- Clear magazine cargo while box is burning

Currently the box will burn for 60 seconds hardcoded, this is to allow time for the ammunition to cook off (since boxes sink into the ground and dissapear when destroyed). Perhaps we can implement a way to burn until all ammo is expended.

* Improve ammo cookoff

* Integrate ammo cookoff with the incendiary grenade

* Disable ammo cook off underwater

* Optimize fnc_detonateAmmunition

I say optimize, the only real performance optimization is using `vectorMultiply`. The rest is readability optimization though!

* Improve ammo box cook off

- Remove unnecessary light source (fire particles provide lighting)
- Add randomness to cook off time
- Cook off begins with fire effect rather than smoke

* Add tracer induced ammo box cook off

Due to limitations in the way arma handles tracer rounds (there's no way to check if an individual projectile is a tracer), only magazines with a high enough tracer density (at least 1 in 4) can cause cook off this way. However this is deemed an acceptable approximation since the chance of this happening should be quite low anyway.

* Decrease amount of explosions from ammo cookoff

* Add is local check for remote event