Add some weapon specific data, change 6.5 cased v
Adds weapon specific data for the MXM, QBU-88, and Noreen Bad News
rifles. The muzzle velocities are the only real changes, but are
important changes for the marksman rifles.

Changed the initSpeed for the 6.5mm Cased ammo that the QBU-88
uses from 820 down to 780, more in-line with the 6.5 caseless that
the MX uses.

The velocity of the QBU has also been changed. The barrel on the
QBU-88 is quite long (25in, as stated in the config), with the base
speed for a 6.5mm round set for a 16in barrel (MX SW). Based on some
quick research on the differences in muzzle velocity of 6.5mm
Creedmoor shot from different barrels[1], I came up with the 1.07
number in this change, as opposed to the reduced velocity of 0.98
previously.

I have also added data for the 6.5mm Lapua and Creedmoor rounds to
the ATragMX table for the QBU that are currently not added.

[1]: http://rifleshooter.com/2016/02/6-5-creedmoor-effect-of-barrel-length-on-velocity-cutting-up-a-creedmoor/