French loc improvement
(French) text as it is clutters the interface interface, making it
uneasy for medic to use the it smoothly.
https://i.gyazo.com/8d7a5bb2f15dd1332cec36f2eef510c8.png

Hopefully suggested changes fixes that.

There were a few typos and, in my opinion, unaccurate or missing
translation.
Based on French Army Field Manual, actual paramedical personnel and past
work on XMedSys.