Improve Japanese translation (#5133)
* Add the Japanese translation for cargo

Add the Japanese translation for cargo, fix to html tag and few improvement

* fix the Japanese transltion

Fix the Japanese translation for attach. it was used wrong html tag.

* changed The Japanese translation

Changed the Japanese translation of azimuth angle

* fix The Japanese translation

Fix the Japanese translation for chemlights. it was used wrong html tag.

* fix The Japanese translation

Fix the Japanese translation for laserpointer. it was used wrong html tag.

* Fix the Japanese translation

Fix the Japanese translation for medical_blood. it was used wrong translation.

* Fix the Japanese translation

Fix the Japanese translation for rearm. it was used wrong translation.

* add The Japanese translation

add The Japanese translation to new words

* Improve The Japanese translation

Improve, fix and change The Japanese translation

* minor fix to Japanese translation

minor fix to Japanese translation. see b1b07c5