Add support for N-bangs (#6848)
* Add support for multi-bang flashbangs

It also adds some randomness to 1-bangs as a side effect,
but I don't think that's a big deal.

Also adds flashbang documentation to wiki.

* Increase rate at which flashbang attenuates

Apologies for the magic numbers. I just tested a few out
until I was happy with the result. The radius could probably
afford to be reduced a little further.

* use max deviation in config, fix wiki file name