Verbose names and No4 Bird
- Change classnames and string keys to be more verbose
- Change No7 Birdshot to No4 Birdshot
While No7 is one of the most common loads for target shooting and hunting small birds the extreme number of pellets is just too hard on FPS.
No4 is not typically used in target shooting, but it is a common load for hunting various birds so I think it's realistic to have it for civlian purposes and the number of pellets is substatially less so the FPS hit isn't so bad.
If someone else could test the No4 to see how it feels to them I'd appreciate it.