Add AGM-114L Active Radar Homing Hellfire (#7337)
* Create AGM-114L

* If we lose LOS dont continue tracking magically Allow for datalinked targets to donate data to the missile.

* Update documentation

* RHS 2x hellfire compatability

* Make ARH more realistic by not allowing to switch targets after firing

* Fix filename. Change Hellfire attack profile to lead target. Switch to two LOS checks.

We check two Line's of Sight to ensure that we are 100% gone from the target. A raw LOS check will be blocked by bushes and light trees while the checkVisibility wont while on the otherhand smoke will block a visiblity check but not a raw LOS check. We get best of both worlds with this.
I changed the attack profile so that it will lead moving targets. This isnt needed with the laser version because the user will "lead" the target if needed, but with the radar scan we have velocity information so we might as well lead the target as much as possible

* Change function calls to FUNC macro. Slightly change radar logic. Up poll frequency to 7hz

Instead of the missile immediately going active when the shooter doesn't have radar, check if targets are in datalink. If they are, use the datalink to guide the missile instead of its internal radar.

* Add logic for missiles launched without target

If a missile is fired without a locked target, it will immediately go active and target the first thing its seeker picks up. This is an incredibly dangerous trait of active radar homing missiles and is so in this implementation. Be careful!

* Change from `exitWith` to basic `then`

Legacy code that never got changed. This is essentially what happened before

* Update CfgMagazineWells.hpp

Co-authored-by: PabstMirror <pabstmirror@gmail.com>